split backup package to two files: init, manager
